Mudança de tipo de RFID para long Não queremos lidar com valores negativos, mas o tipo uint não é cls-compliant, então usaremos o long no servidor

+++ src_server/VendingMachine.Core/Model/ClientCard.cs
@@ -10,7 +10,7 @@ namespace VendingMachine.Core.Model
     public class ClientCard
     {
         [Key]
-        public virtual int Rfid { get; set; }
+        public virtual long Rfid { get; set; }
 
         [Display(Name ="Apelido")]
         public virtual string Alias { get; set; }
